backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
pydantic==2.5.3
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.3
faiss-cpu==1.7.4
tiktoken==0.5.2
openai==1.10.0
pillow==10.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytesseract==0.3.10
bcrypt==4.1.2
psutil==5.9.8
uvloop==0.19.0
httptools==0.6.1
